Split out event args cache and lookup
When we generated the EventArgsCache, we might get property names which
differ only in special characters, which need to be escaped. This means
that we might get collisions on the generated property names on the
EventArgsCache.

This previously meant that we needed to generate a mapping from all
property names -> the generated property name on the EventArgsCache class,
which was used in generating each partial type. Whenever this mapping
changed, all partial types would have to be re-generated.

Now we just store a list of collisions, and assume that if a property name
isn't in the list of collisions, then the default escaping mechanism was
fine and didn't generate a collision. This means that we only need to
re-generate all partial types of the list of collisions changes, which
should be significantly less often.